+++ ref/acceptance_validation_test.rb
