+++ ref/attribute_test.rb
