+++ ref/changed_attribute_test.rb
