+++ ref/confirmation_validation_test.rb
