+++ ref/exclusion_validation_test.rb
