+++ ref/format_validation_test.rb
