+++ ref/inheritance_test.rb
