+++ ref/integration/hanami_entity_compatibility_test.rb
