+++ ref/size_validation_test.rb
