+++ ref/test_helper.rb
@@ -1,0 +1,29 @@
+require 'rubygems'
+require 'bundler/setup'
+
+if ENV['COVERAGE'] == 'true'
+  require 'simplecov'
+  require 'coveralls'
+
+  SimpleCov.formatters = [
+    SimpleCov::Formatter::HTMLFormatter,
+    Coveralls::SimpleCov::Formatter
+  ]
+
+  SimpleCov.start do
+    command_name 'test'
+    add_filter   'test'
+  end
+end
+
+require 'minitest/autorun'
+$:.unshift 'lib'
+require 'hanami/validations'
+
+module Hanami::Validations::ValidationIntrospection
+  def defined_validation?(name)
+    validations.instance_variable_get(:@validations).keys.include?(name)
+  end
+end
+
+require 'fixtures'
