+++ ref/validations/coercions_test.rb
