+++ test/test_helper.rb
@@ -1,29 +1,11 @@
 require 'rubygems'
 require 'bundler/setup'
 
-if ENV['COVERAGE'] == 'true'
-  require 'simplecov'
+if ENV['COVERALL']
   require 'coveralls'
-
-  SimpleCov.formatters = [
-    SimpleCov::Formatter::HTMLFormatter,
-    Coveralls::SimpleCov::Formatter
-  ]
-
-  SimpleCov.start do
-    command_name 'test'
-    add_filter   'test'
-  end
+  Coveralls.wear!
 end
 
 require 'minitest/autorun'
-$:.unshift 'lib'
+$LOAD_PATH.unshift 'lib'
 require 'hanami/validations'
-
-module Hanami::Validations::ValidationIntrospection
-  def defined_validation?(name)
-    validations.instance_variable_get(:@validations).keys.include?(name)
-  end
-end
-
-require 'fixtures'
